requirements.txt
nltk==3.8.1
scikit-learn==1.3.0
flask==2.3.3
streamlit==1.25.0
pandas==2.0.3
numpy==1.24.3
gensim==4.3.1
whoosh==2.7.4
python-dotenv==1.0.0
flask-cors==4.0.0